chore!: remove outdated constants and replace with extracted enum
