+++ backend/app/api/v1/repositories/refresh_jobs.py
@@ -2,8 +2,7 @@ from datetime import UTC, datetime
 
 from sqlalchemy.orm import Session
 
-from app.constants.jobs import JobName
-from app.models.enums import JobStatus
+from app.models.enums import JobName, JobStatus
 from app.models.refresh_job import RefreshJob
